+++ src/index.ts
@@ -1,7 +1,7 @@
 import './types/meteor/ModuleAugmentation';
 
+export * from './guards';
 export * from './Definitions';
-export * from './Guard';
 export * from './MeteorTypeValidation';
 
 export type * from './types/ValidatedResources';