+++ tests/type-tests/guard-contexts.ts
@@ -1,39 +1,7 @@
-import {
-    defineMethods,
-    definePublications, Guard,
-    type UnwrapMethods,
-    type UnwrapPublications, type WrappedContext,
-} from '../../src';
 import * as v from 'valibot';
-
-class UserAuthenticated extends Guard {
-    public validate(): asserts this is { context: { userId: string } } {}
-    public get validatedContext() {
-        this.validate();
-        return this.context;
-    }
-}
-
-class CreatedByCurrentUser extends UserAuthenticated {
-    constructor(context: WrappedContext, protected readonly params: [{ createdBy: string }, ...rest: unknown[]]) {
-        super(context, params);
-    }
-    
-    public validate(): asserts this is {
-        context: { userId: string },
-        params: [{ createdBy: string }, ...unknown[]]
-    } {
-        if (!this.params[0]?.createdBy) {
-            throw new Error('No user ID')
-        }
-        return super.validate();
-    }
-    
-    public get validatedContext() {
-        this.validate();
-        return this.context;
-    }
-}
+import { defineMethods, definePublications, exposeMethods, exposePublications } from '../../src';
+import { UserAuthenticated } from '../../src/guards/UserAuthenticated';
+import { CreatedByCurrentUser } from '../lib/CreatedByCurrentUserGuard';
 
 export const AllMethods = defineMethods({
     'user:todo.add': {
@@ -85,11 +53,16 @@ export const AllPublications = definePublications({
             entry._id
         }
     }
-})
+});
+
+const Methods = exposeMethods(AllMethods);
+const Publications = exposePublications(AllPublications);
+type Publications = typeof Publications;
+type Methods = typeof Methods;
 
 declare module 'meteor/meteor' {
-    interface DefinedPublications extends UnwrapPublications<typeof AllPublications> {}
-    interface DefinedMethods extends UnwrapMethods<typeof AllMethods> {}
+    interface DefinedPublications extends Publications {}
+    interface DefinedMethods extends Methods {}
 }